pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scadable-sdk"
version = "0.2.0"
description = "Scadable Edge SDK — write device logic in Python, compile to native"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [{ name = "Scadable" }]
keywords = ["iot", "scada", "industrial", "edge", "modbus"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: System :: Hardware",
]
dependencies = [
    "typer[all]>=0.12",
    "toml>=0.10",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.urls]
Homepage      = "https://scadable.com"
Documentation = "https://docs.scadable.com"
Repository    = "https://github.com/scadable/scadable-sdk"
Changelog     = "https://github.com/scadable/scadable-sdk/blob/main/CHANGELOG.md"

[project.scripts]
scadable = "scadable.cli.main:app"

[project.optional-dependencies]
test = [
    "pytest>=8",
    "pytest-cov>=5",
    "pytest-snapshot>=0.9",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-pyyaml>=6.0",
    "types-toml>=0.10",
]

[tool.setuptools.packages.find]
include = ["scadable*"]
exclude = ["tests*", "examples*", "docs*"]

[tool.setuptools.package-data]
scadable = ["py.typed"]

[tool.pytest.ini_options]
testpaths   = ["tests"]
addopts     = "-q --strict-markers"
xfail_strict = true

[tool.coverage.run]
source = ["scadable"]
omit   = ["scadable/cli/templates/*"]

[tool.coverage.report]
fail_under = 70
show_missing = true
skip_covered = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle / pyflakes
    "I",             # isort
    "UP",            # pyupgrade
    "B",             # bugbear
    "SIM",           # simplify
]
ignore = ["E501"]    # line length handled by formatter

[tool.mypy]
python_version       = "3.10"
warn_unused_ignores  = true
no_implicit_optional = true
ignore_missing_imports = true
files                = ["scadable"]

[[tool.mypy.overrides]]
module = ["toml.*", "yaml.*"]
ignore_missing_imports = true
